+++ src/cz_cpp/DomainInfo.h
@@ -6,7 +6,7 @@
 #
 # CubeZ
 #
-# Copyright (C) 2018 Research Institute for Information Technology(RIIT), Kyushu University.
+# Copyright (C) 2018-2020 Research Institute for Information Technology(RIIT), Kyushu University.
 # All rights reserved.
 #
 ###################################################################################